+++ relay.h
@@ -40,6 +40,7 @@ extern unsigned char mode;
 #endif
 
 extern char relay_hostname[];
+extern char noexpire;
 
 enum logdst { LOGOUT, LOGERR };